Create http_server adapater
Remove http -listen functionality and place it in its own adapter called
`http_server`
