+++ lib/runtime/adapters.js
@@ -8,7 +8,7 @@ var errors = require('../errors');
 var Module = require('module');
 var logger = require('../logger').getLogger('juttle-adapter');
 
-var BUILTIN_ADAPTERS = ['file', 'http', 'stdio', 'stochastic'];
+var BUILTIN_ADAPTERS = ['file', 'http', 'stdio', 'stochastic', 'http_server'];
 
 var adapters = {};
 var adapterConfig = {};
